scripts: support prune, storage config;

+++ scripts/clusterup.sh
@@ -11,9 +11,9 @@ bnbHolderAddr=''
 function exit_previous() {
 	# stop client
 	echo "kill all nodes!"
-	ps -ef  | grep geth | grep  bsc-deploy| awk '{print $2}' | xargs kill
-	ps -ef  | grep bootnode | grep  bsc-deploy| awk '{print $2}' | xargs kill
-	sleep 20
+	ps -ef  | grep ${workspace}/bin/geth | awk '{print $2}' | xargs kill
+	ps -ef  | grep ${workspace}/bin/bootnode | awk '{print $2}' | xargs kill
+  sleep 30
 }
 
 function generate_static_peers() {
@@ -58,6 +58,10 @@ function prepare() {
          echo "bin/geth do not exist!"
          exit 1
     fi
+    if ! [[ -f ${workspace}/bin/bootnode ]];then
+         echo "bin/bootnode do not exist!"
+         exit 1
+    fi
     rm -rf ${workspace}/clusterNode
     cd  ${workspace}/genesis
     git stash
@@ -74,6 +78,14 @@ function prepareGethEnv(){
         mkdir -p ${workspace}/clusterNode/node${i}
         echo 'password' >> ${workspace}/clusterNode/password.txt
         ${workspace}/bin/geth --datadir ${workspace}/clusterNode/node${i} account new --password ${workspace}/clusterNode/password.txt > ${workspace}/clusterNode/validator${i}Info
+        validatorAddr=("${validatorAddr[@]}" `cat ${workspace}/clusterNode/validator${i}Info|grep 'Public address of the key'|awk '{print $6}'` )
+        validatorSecretLoc=("${validatorSecretLoc[@]}" `cat ${workspace}/clusterNode/validator${i}Info|grep  'Path of the secret key file'|awk '{print $7}'`)
+    done
+}
+
+function loadGethEnv(){
+    num=$1
+    for((i=1;i<=$num;i++)); do
         validatorAddr=("${validatorAddr[@]}" `cat ${workspace}/clusterNode/validator${i}Info|grep 'Public address of the key'|awk '{print $6}'` )
         validatorSecretLoc=("${validatorSecretLoc[@]}" `cat ${workspace}/clusterNode/validator${i}Info|grep  'Path of the secret key file'|awk '{print $7}'`)
     done
@@ -115,29 +127,88 @@ function generateGenesis(){
     done
 }
 
-function startNode() {
+function startFullNodeWithExpiry() {
     num=$1
+    nodeNum=$2
+    remote=$3
     for((i=1;i<=$num;i++)); do
-        validatorIndex=$(($i-1))
-        nohup ${workspace}/bin/geth -unlock ${validatorAddr[$validatorIndex]} --password "${workspace}/clusterNode/password.txt" --mine \
-         --rpc.allow-unprotected-txs --light.serve 50 \
-         --gcmode archive --ws --datadir ${workspace}/clusterNode/node${i} --config ${workspace}/clusterNode/node${i}/config.toml \
-         --metrics --pprof --pprof.port "$((6060+$i))" --http.corsdomain "https://remix.ethereum.org" > ${workspace}/clusterNode/node${i}/bsc-node.log 2>&1 &
-        
-        echo "start validator $i, ${validatorAddr[$validatorIndex]}"
-        echo ${validatorAddr[$validatorIndex]} > ${workspace}/clusterNode/validator${i}addr
+        validatorIndex=$(($nodeNum-1))
+        nohup ${workspace}/bin/geth -unlock ${validatorAddr[$validatorIndex]} --http --http.port "$((8501+$nodeNum))" --ws.port "$((8545+$nodeNum))" \
+         --config ${workspace}/clusterNode/node${nodeNum}/config.toml \
+         --authrpc.port "$((8550+$nodeNum))" --password "${workspace}/clusterNode/password.txt" \
+         --mine --miner.etherbase ${validatorAddr[$validatorIndex]} --rpc.allow-unprotected-txs --allow-insecure-unlock --light.serve 50 \
+         --gcmode full --ws --datadir ${workspace}/clusterNode/node${nodeNum} \
+         --metrics  --metrics.addr "0.0.0.0" --metrics.port "$((6060+$nodeNum))" --pprof --pprof.port "$((6070+$nodeNum))" --http.corsdomain "*" --rpc.txfeecap 0 \
+         --state-expiry --state-expiry.remote ${remote} > ${workspace}/clusterNode/node${nodeNum}/geth-$(date +"%Y%m%d_%H%M").log 2>&1 &
+
+        echo "start validator $nodeNum, enable state expiry, miner: ${validatorAddr[$validatorIndex]}"
+        nodeNum=$(($nodeNum+1))
+
+        sleep 1
     done
 }
 
-function restartNode() {
+function startFullNodeNoExpiry() {
     num=$1
+    nodeNum=$2
     for((i=1;i<=$num;i++)); do
-        nohup ${workspace}/bin/geth -unlock $(cat ${workspace}/clusterNode/validator${i}addr) --password "${workspace}/clusterNode/password.txt" --mine \
-         --rpc.allow-unprotected-txs --light.serve 50 \
-         --gcmode archive --ws --datadir ${workspace}/clusterNode/node${i} --config ${workspace}/clusterNode/node${i}/config.toml \
-         --metrics --pprof --pprof.port "$((6060+$i))" --http.corsdomain "https://remix.ethereum.org" > ${workspace}/clusterNode/node${i}/bsc-node.log 2>&1 &
+        validatorIndex=$(($nodeNum-1))
+        nohup ${workspace}/bin/geth -unlock ${validatorAddr[$validatorIndex]} --http --http.port "$((8501+$nodeNum))" --ws.port "$((8545+$nodeNum))" \
+         --config ${workspace}/clusterNode/node${nodeNum}/config.toml \
+         --authrpc.port "$((8550+$nodeNum))" --password "${workspace}/clusterNode/password.txt" \
+         --mine --miner.etherbase ${validatorAddr[$validatorIndex]} --rpc.allow-unprotected-txs --allow-insecure-unlock --light.serve 50 \
+         --gcmode full --ws --datadir ${workspace}/clusterNode/node${nodeNum} --rpc.txfeecap 0 \
+         --metrics  --metrics.addr "0.0.0.0" --metrics.port "$((6060+$nodeNum))" --pprof --pprof.port "$((6070+$nodeNum))" --http.corsdomain "*" > ${workspace}/clusterNode/node${nodeNum}/geth-$(date +"%Y%m%d_%H%M").log 2>&1 &
 
-        echo "restart validator $i, $(cat ${workspace}/clusterNode/validator${i}addr)"
+        echo "start validator $nodeNum as full node, miner: ${validatorAddr[$validatorIndex]}"
+        nodeNum=$(($nodeNum+1))
+        sleep 1
+    done
+}
+
+function pruneFullNodeNoExpiry() {
+    num=$1
+    nodeNum=$2
+    for((i=1;i<=$num;i++)); do
+        validatorIndex=$(($nodeNum-1))
+        nohup ${workspace}/bin/geth snapshot prune-state --config ${workspace}/clusterNode/node${nodeNum}/config.toml \
+         --datadir ${workspace}/clusterNode/node${nodeNum} > ${workspace}/clusterNode/node${nodeNum}/geth-prune-$(date +"%Y%m%d_%H%M").log 2>&1 &
+
+        echo "start prune validator $nodeNum as full node"
+        nodeNum=$(($nodeNum+1))
+        sleep 1
+    done
+}
+
+function pruneFullNodeWithExpiry() {
+    num=$1
+    nodeNum=$2
+    for((i=1;i<=$num;i++)); do
+        validatorIndex=$(($nodeNum-1))
+        nohup ${workspace}/bin/geth snapshot prune-state --config ${workspace}/clusterNode/node${nodeNum}/config.toml \
+         --datadir ${workspace}/clusterNode/node${nodeNum} \
+         --state-expiry > ${workspace}/clusterNode/node${nodeNum}/geth-prune-$(date +"%Y%m%d_%H%M").log 2>&1 &
+
+        echo "start prune validator $nodeNum as full node, enable state expiry feature"
+        nodeNum=$(($nodeNum+1))
+
+        sleep 1
+    done
+}
+
+function storageProfile() {
+    num=$1
+    nodeNum=$2
+    for((i=1;i<=$num;i++)); do
+        validatorIndex=$(($nodeNum-1))
+        echo "start profile validator $nodeNum"
+        ${workspace}/bin/geth db inspect \
+         --datadir ${workspace}/clusterNode/node${nodeNum} > ${workspace}/clusterNode/node${nodeNum}/geth-storage-inspect-$(date +"%Y%m%d_%H%M").log 2>&1
+        ${workspace}/bin/geth db inspect-trie \
+         --datadir ${workspace}/clusterNode/node${nodeNum} latest 1000 > ${workspace}/clusterNode/node${nodeNum}/geth-trie-inspect-$(date +"%Y%m%d_%H%M").log 2>&1
+
+        nodeNum=$(($nodeNum+1))
+        sleep 1
     done
 }
 
@@ -146,14 +217,19 @@ CMD=$1
 case ${CMD} in
 start)
     exit_previous
-    validatorNum=3
-    if [ ! -z $3 ] && [ "$3" -gt "0" ]; then
-      validatorNum=$3
-    fi
     bnbHolderAddr="0xD8C0Aa483406A1891E5e03B21F2bc01379fc3b20"
     if [ ! -z $2 ] && [ "$2" -gt "0" ]; then
       bnbHolderAddr=$2
     fi
+    fullNumWithExpiry=2
+    if [ ! -z $3 ] && [ "$3" -gt "0" ]; then
+      fullNumWithExpiry=$3
+    fi
+    fullNumNoExpiry=1
+    if [ ! -z $4 ] && [ "$4" -gt "0" ]; then
+      fullNumNoExpiry=$4
+    fi
+    validatorNum=fullNumWithExpiry+fullNumNoExpiry
     echo "===== generate node key ===="
     generate_nodekey $validatorNum
     echo "===== preparing ===="
@@ -161,19 +237,64 @@ start)
     prepareGethEnv $validatorNum
     generateGenesis $validatorNum
     echo "===== starting client ===="
-    startNode $validatorNum
-    ps -ax | grep geth
+    startFullNodeNoExpiry fullNumNoExpiry 1 # By default, last node is remoteDB
+    remote="http://127.0.0.1:$((8501+1))"
+    startFullNodeWithExpiry fullNumWithExpiry $((fullNumNoExpiry+1)) $remote
     echo "Finish deploy"
     ;;
 restart)
     exit_previous
-    validatorNum=3
+    fullNumWithExpiry=2
     if [ ! -z $3 ] && [ "$3" -gt "0" ]; then
-      validatorNum=$3
+      fullNumWithExpiry=$3
     fi
+    fullNumNoExpiry=1
+    if [ ! -z $4 ] && [ "$4" -gt "0" ]; then
+      fullNumNoExpiry=$4
+    fi
+    validatorNum=fullNumWithExpiry+fullNumNoExpiry
+    echo "===== preparing ===="
+    loadGethEnv $validatorNum
     echo "===== restarting client ===="
-    restartNode $validatorNum
+    startFullNodeNoExpiry fullNumNoExpiry 1 # By default, last node is remoteDB
+    remote="http://127.0.0.1:$((8501+1))"
+    startFullNodeWithExpiry fullNumWithExpiry $((fullNumNoExpiry+1)) $remote
     echo "Finish restart"
+    ;;
+prune)
+    exit_previous
+    fullNumWithExpiry=2
+    if [ ! -z $3 ] && [ "$3" -gt "0" ]; then
+      fullNumWithExpiry=$3
+    fi
+    fullNumNoExpiry=1
+    if [ ! -z $4 ] && [ "$4" -gt "0" ]; then
+      fullNumNoExpiry=$4
+    fi
+    validatorNum=fullNumWithExpiry+fullNumNoExpiry
+    echo "===== preparing ===="
+    loadGethEnv $validatorNum
+    echo "===== restarting client ===="
+    pruneFullNodeNoExpiry fullNumNoExpiry 1
+    pruneFullNodeWithExpiry fullNumWithExpiry $((fullNumNoExpiry+1))
+    echo "Finish prune"
+    ;;
+storage)
+    exit_previous
+    fullNumWithExpiry=2
+    if [ ! -z $3 ] && [ "$3" -gt "0" ]; then
+      fullNumWithExpiry=$3
+    fi
+    fullNumNoExpiry=1
+    if [ ! -z $4 ] && [ "$4" -gt "0" ]; then
+      fullNumNoExpiry=$4
+    fi
+    validatorNum=fullNumWithExpiry+fullNumNoExpiry
+    echo "===== preparing ===="
+    loadGethEnv $validatorNum
+    echo "===== restarting client ===="
+    storageProfile validatorNum 1
+    echo "Finish storage profile"
     ;;
 stop)
     echo "===== stopping client ===="
@@ -181,7 +302,6 @@ stop)
     echo "===== client stopped ===="
     ;;
 *)
-    echo "Usage: localup.sh start | restart | stop | genreate \${number} "
+    echo "Usage: clusterup.sh start | stop"
     ;;
 esac
-
